requirements.txt
PyYAML==3.10
easywebdav==1.0.7
requests==1.2.3
wsgiref==0.1.2
